Unit test cleanup.
Moved all unittests from script.py into their own modules in the tests
directory where nosetests can find them.

The tests no longer use tool/main.py to start and stop dispersy,
instead this is done using setUpClass and tearDownClass in
dispersytestclass.py.
